refactor: split CLI wiring from runCLI to enable end-to-end tests
src/cli.ts called program.parse() at module scope and kept runCLI
module-private, so importing the module to test it would run the CLI
- there was no way to exercise file output modes, config discovery,
--dry-run, or option validation without shelling out.

Move the behavior into src/cli-runner.ts, exporting runCLI and
CLIOptions. src/cli.ts now only declares commander options and calls
runCLI from its action handler.

Fixes #392
